rename module name for _constrained_dominates
